refactor(api): consolidate the four original-map lookup implementations into a shared originalMapLateral subquery
